Create LICENSE
This commit adds the MIT License to the repository, granting users the freedom to use, modify, and distribute the code under permissive terms. By including the MIT License, we aim to foster collaboration and encourage contributions from the open-source community. The MIT License is a widely recognized and respected license in the software industry, ensuring that our project remains open and accessible to all. Contributors and users of this repository can now confidently utilize the codebase while adhering to the terms specified in the license.